Create 88-merge-sorted-array.py
88 https://www.loom.com/share/6eb54264652844f28942d1584fd55455?sid=ac700682-0a50-4133-a17c-7a7ca1eba5cf

We are given two sorted arrays and m and n, which represent the number of elements in the two arrays. We want to combine them into one array in non-decreasing order.
The key idea is to use a three pointer system so that we can swap the elements in place without using extra memory and efficient time.