Create 27-remove-element.py
27
https://www.loom.com/share/4f325b255ec843e0966f35b8bf109b85?sid=ecb9a6e0-c164-468f-9d5a-46170fadc647 

We have an array and an integer value. We want to remove all instances of this value in the array.
The key idea is to use a counter for k and swap places with the original array and k, so that only elements that are not the integer val are included in k. We then return k and all instances of val are now removed.