Create 26_Remove_Duplicate_From-Sorted_Array.py
26.

https://www.loom.com/share/2accd3094141403399939f0a6f3b3070?sid=15b3a660-4e30-4c25-9a19-6c84f2725a77 

We have an array, nums, and we want only the unique elements in the array, so no duplicates should be found in the array after running the code.
The key idea is to use a two pointer system, i and j, where i starts at 0 and j starts at 1. We compare i and j and if they are the same we do nothing, but if they are not the same, we increment i and swap i and j until we reach the end of the array and then we return i + 1 to get the number of unique elements.