Create 121_Best_Time_To_Buy_And_Sell_Stock.py
121.

https://www.loom.com/share/102414d79d894c71a8b6b61f175f9af1?sid=72ebedb6-7be5-4ef1-ae3d-b035fa100338 

We have an array of prices and we want to know what our max profit will be. We must choose one day to buy and another to sell. If we cannot achieve any profit then we return 0.
We need to find the max profit from buying and selling on different days, where the buying day comes before the selling day. We go through the prices once, keeping track of the min price encountered so far and the max profit possible as we proceed.