Create 14_Longest_Common_Prefix.py
14. 

Need to find the longest common prefix string amongst an array of strings.
Use the first string as the prefix to compare to the other strings. We start with the last letter in the string and continue to shorten it until it matches with the prefix. If the new string gets to the end and it never matches, then the prefix return an empty string.