Create 125_Valid_Palindrome.py
125.

The key idea is to use two pointers. One starts at the first position, the second starts at the last position. If the two letters match then continue to move the two pointers toward the center. If at any point the two positions don’t match, then return False, but if we get to the center, then return True. We also can convert all characters to lowercase and only account for alpha numeric characters when iterating through the string.