Create 392_Is_Subsequence.py
392.

We need to find if the string s is a subsequence of t
Use two pointers for s and t. If s equals t then we increment s, otherwise just increment t. If we get to the end and all characters in s have been found then s is a subsequence of t.