Create 1_Two_Sum.py
1.

Use a hashmap to store the number and its index. If we have found a complement, which is the target minus the number, in the hashmap, then we return the indices at which they were found, if not then we store the number, index in the hash map and continue to iterate through nums until we do.