Create 20_Valid_Parentheses.py
20.

Use a stack, where we iterate through our string, and if it is open, we add it to our stack. If it is closed we check to see if the stack is empty and if so we return false, if not then we check if the top element has a matching opening bracket and we then pop it from the stack if it does. If it does not then we return False. After iterating through s, if the stack is empty, we return true, if not, return false, as not all matches were found.