Create 141_Linked_List_Cycle.py
141.

We are trying to find if there exists a cycle in a linked list. A cycle exists if a node in the list can be reached by continuously following the next pointer.
The key idea is to use a slow and fast pointer. The slow pointer will move one position at a time, the fast pointer will move two positions at a time. If they ever meet at the same position, then there is a cycle and we return true, if not possible for them to meet, return false.