Create 21_Merge_Two_Sorted_Lists.py
21.

We are given the heads of two linked lists, we want to merge the two lists together by splicing the first node of each list.
We can iterate through both lists simultaneously and store them in a new list called dummy. We always add the smaller node first. If there are any remaining nodes after traversing we add them to the dummy list. We then return the next node of dummy, which is the head of the new list.