Create 104_Maximum_Depth_Of_Binary_Tree.py
104.

https://leetcode.com/problems/maximum-depth-of-binary-tree/description/

We are given the root of a binary tree and want to find its maximum depth.
The key idea is to use recursion to go as deep as possible in the tree and use max to find which depth is greater out of the left and right. We add 1 to the answer to get the correct height.