Create 100_Same_Tree
100. 

https://www.loom.com/share/f513d1a8503c4d399ed99e73e85fc730?sid=477c7fd1-d905-4e67-b512-d6efa700b287

We are given the roots of trees named p and q. We need to determine if they are the same tree or not.
We can do this by checking if both trees are empty, check if one is empty and the other is not, and check if the values for both trees are the same and the left and right heights of the tree are the same.