Create 70_Climbing_Stairs.py
70.

The key idea is to use a variation of the fibonacci sequence using dynamic programming. We know that there are two ways to take steps, either one at a time or two at a time, so we make that our base case. Then, we calculate ways for each step from 3 up to n. Update steps to reflect the current step, then return the number of ways to take steps.