Create 217_Contains_Duplicate.py
217.

We are given an integer array of nums, and if there are duplicate elements in the array then return true.
The key idea is to use a set to store seen elements. If a duplicate is found in the set, then we return true, if we get to the end of the array and no duplicates found, then we return false.