Create 347_Top_K_Frequent_Elements.py
347.

https://www.loom.com/share/3a904a4b74f149f2bbaf3d9bb70fb744?sid=ed6549ab-cb44-4894-ab0f-28c1635fc6d8

The key idea is to use a heap to store at most the top k elements. We can use a counter to store frequencies and then push a tuple of the number and the frequency onto the heap. If the heap is bigger than k, then pop from the heap until it is not. We then extract the number from the heap to get the result.