Create 238_Product_Of_Array_Except_Self.py
238.

The key idea to solving this is passing over the array twice. In the first pass, we create an array of 1's and accumulate products of elements to the left. Then we pass from the right and accumulate products of elements to the right by using the left pass resulting array.