Create 128_Longest_Consecutive_Sequence.py
128.

https://www.loom.com/share/8795d9546fac4ea699efc1459b80d5fa?sid=b57f9835-cdf8-4759-8aff-a937eddffcfb

Given an unsorted array of integers nums, return the length of the longest consecutive elements sequence. You must write an algorithm that runs in O(n) time.
The key idea is to use a set, so that you can process each number once in O(1) operations. We will keep track of the longest consecutive sequence using the max function.