Create 167_Two_Sum_II_Input_Array_Is_Sorted.py
167.

Given a 1-indexed array of integer numbers that is already sorted in non-decreasing order, find two numbers such that they add up to a specific target number. Let these two numbers be numbers[index1] and numbers[index2] where 1 <= index1 < index2 <= numbers.length. Return the indices of the two numbers, index1 and index2, added by one as an integer array [index1, index2] of length 2. The tests are generated such that there is exactly one solution. You may not use the same element twice. Your solution must use only constant extra space.

We can perform a binary search where we have a left and right pointer. The left at the beginning of the array, the right at the end of it. We check if the sum of these two points equals the target and if so return the left and right index + 1, since it is 1-indexed. If it is less than the target move left, if it is greater move the right pointer.