Create 5_Longest_Palindromic_Substring.py
5.

The expand around center solution works by considering each position in the string as a potential center of a palindrome, where for each position we consider both odd-length palindromes (with the center at current position) and even-length palindromes (with the center between current and next position). For each center position, we expand outwards as long as the characters on both sides match, keeping track of the longest palindrome we've found. Finally, we return the substring corresponding to the longest palindrome found throughout this process.