Create 322_Coin_Change.py
322.

To solve the coin change problem, we use dynamic programming by creating an array where each position represents the minimum coins needed to make that amount. For each amount from 1 to target, we try using each available coin and check if we can get a better solution by using that coin plus the minimum coins needed for the remaining amount. Once we've processed all amounts up to our target, we check if we found a valid solution (if not, we return -1) and return the minimum number of coins needed.