Create 62_Unique_Paths.py
62.

Start by creating a grid where each cell will store the number of ways to reach that cell, initializing the first row and first column with 1s since there's only one way to reach those cells (by moving only right or only down respectively). For each remaining cell, calculate the number of ways to reach it by adding the number of ways to reach the cell above it and the number of ways to reach the cell to its left. Finally, the value in the bottom-right cell of your grid will give you the total number of unique paths from start to finish.