Create 1143_Longest_Common_Subsequence.py
1143.

Create a 2D array where each cell [i][j] represents the longest common subsequence length between text1[0...i-1] and text2[0...j-1]. For each cell, if the characters at the current position match (text1[i-1] == text2[j-1]), take the value from the diagonal cell (top-left) and add 1; if they don't match, take the maximum value from either the cell above or the cell to the left. The final answer will be in the bottom-right cell of the array, representing the length of the longest common subsequence between the entire strings.