Create 1570_Dot_Product_Of_Two_Sparse_Vectors.py
1570.

we create a dictionary that only stores non-zero elements,
with indices as keys and elements as values.

we iterate over the two vectors and check if the smaller
dictionary has the same indices in the other vector. If so, 
we sum the product of the matching values. We also want to swap vectors in the case where one is smaller than the other, so that we always iterate over the smallest vector leading to more efficient computations.