Create 1091_Shortest_Path_In_Binary_Matrix.py
1091.

We can use BFS(Breadth-First Search) for this problem as it will explore all possible paths and return the shortest path first, since each step has equal weight. For each cell, check all 8 possible directions to find valid cells, once visited mark the cell as so and increase the length for each cell. If we reach the target cell then return the path length, if not then return -1 as it is not possible.